This is the first working version of the bow-text-feature-module
consists of 2 files. the txt file contains the library and the .py file the code
currently the library hold the top 1000 words for negative classification.
in the .py-file-comments is a script to retrain the library with an arbitrary size.
I decided to limit the size to get some  meaningfull values out of the score function
instead of very, very small values.