Added a feature-module, which detects errors in the feature-extraction process
All other modules have to include a attribute self.error, which is a
boolean and flagged true if the feature extraction by get_function()
crashed or produced invalid values.
The flag-value is resetet by this module.
Other than every other module, this module  needs a reference to every
other module, organized in a list in order to work properly. This
module must be called last.